dbusdummyservice: dbus paths are now writeable
This is handy while testing. Open one terminal to run the dummy-
service. Open another to run the project you are working on and that
you want to test. And if you need to change, for example, battery
state of charge to 30%, open a third terminal and call the SetValue
function over the dbus.