VeDbusItemImport: fix bug, it should always get signals
Problem:
VeDbusItemImport returns the cached value if someone calls
VeDbusItemImport.get_value(..)). Therefore code relies on being subscribed
to the PropertiesChanged signal. Without those signals coming in, get_value()
would forever return the initial value.

The code was trying to be clever, and only connected to the signals when a
callback function was passed to VeDbusItemImport.__init__(). All code that I
can think of does pass the callback, so this bug probably did not occur on
any of the CCGX projects.

Solution:
Took away the clever stuff, and moved the connect-to-signal() call to
__init__().

Also renamed dbus._object to dbus._proxy, that is a better name.