Fixed memleak in VeDbusItemImport, and in dbusmonitor
All projects using dbusmonitor should upgrade to this version of velib_python.
VeDbusItemImport objects were not being released when a D-Bus service
dissapears from the D-Bus, because the Python garbage collector does not remove
objects with a cyclic reference that have a __del__() statement.

For details, see comments in vedbus.py -> VeDbusItemImport.

To reproduce the problem, run vrmlogger, vrmpubnub or just dbusmonitor.py, and
make a script that continuously publishes a service on the D-Bus and goes off-
line again after a few seconds. All this ofcourse with a version of
VeDbusItemImport that has a __del__() statement!