dbusdummyservice: add receiving data and non-updating
Receiving data:
When receiving data (someone else calls SetValue() on an object
that is exported to the D-Bus by the dbusdummyservice), it writes
the received data to stdout.

Non updating values:
You can now leave out the 'update' key in the dictionary. See
also example code on bottom of the script.