Added logger.py: default logging setup
The advantage of the logging config made here is that all INFO and
DEBUG messages go to stdout. And all other messages (WARN, ERROR,
etc) go to stderr. Which is usefull when running functional or unit-
tests that starts up several child processes: you don't want all
INFO or even DEBUG info of all the child processes to come up in the
shell. But if something goes wrong you do want to see it.