dbusmonitor: call valueChangedCallback only on a real change
Before the startup performance improvements of dbusmonitor (ie 4cb07f4
and earlier), subscribing to PropertyChanged was done when processing
a NameOwnerChange.

The current, startup-performance improved version, makes only one subscription
to PropertyChanged signals, right at startup.

So, knowing that some (or all) services also emit a PropertyChanged signal
for each and every path when to go onto the D-Bus, the new situation requires
us to disregard PropertyChanges containing a value that is the same to the
value that we already had: the value retrieved when processing the
NameOwnerChange, ie 'new service found handler'.

For vrmlogger, this change reduces the number of config-changes and data-sends
in the event of a new service coming online on the D-Bus. It is back to the
same volume of v2.08~10 and before.

Relates to: https://github.com/victronenergy/venus/issues/172