dbusmonitor: fix missing unwrap
Though the value stored in the cache was being unwrapped, the value as
passed to the callback was not.

And this caused problem with only one specific class of values in vrmlogger,
see https://git.victronenergy.com/ccgx/dbus_vrm/blob/master/vrmlogger.py#L167

This was overlooked when taking out DbusItemImport. I re-checked its code [1],
to see if I have not overlooked anything else. And all is ok now.

[1] https://github.com/victronenergy/velib_python/blob/master/vedbus.py#L305