Fix mosquitto bridge config concurrency
When different projects using the bridge registrator wrote the config,
the path to the certificate was different. This caused the different
processes to think they need to update the config file, because it's
different, causing repeated Mosquitto restarts.

Venus has the same certificate in /etc, so we can point to that
directly.