dbusmonitor: revert more strict error checking
In recent changes, the logic in error handling was swapped: instead of raising
and aborting on two specific errors, and silently accepting all others, it was
turned around:

Only silently accept one type, and fail on all others.

Which doesn't work. See [1].

So revert back to the old mechanism. There is no reason to be strict.

https://github.com/victronenergy/venus-private/issues/55