Graceful stop in mosquitto_bridge_registrator.py
This is needed for fixing mqtt-rpc and dbus-mqtt. It allows for a more
graceful shutdown, allowing the server response to be used and the
generated password to be saved (instead of the process just dying and
not getting a chance to save it). Otherwise the server may have
processed the POST, but the client then ignores it. This causes
subsequent MQTT password posts to fail, because it will post new random
passwords, and which don't match was already stored, obviously.

Change is backwards compatible.

https://github.com/victronenergy/venus-private/issues/72