Fix lost posted mqtt password
When we would post a password and the server accepts it but we never see
the response (because of a dropped connection for instance), subsequent
posting new random passwords failed, because one was already stored.

So, when we post a password, we're storing it now, so that we post the
same one in case it fails, and then we are sure the server accepts it.

Backwards compatible change (it still reads the password from the
Mosquitto config file, if it's present).

https://github.com/victronenergy/venus-private/issues/72