dbus-monitor: add check if an path is ever seen
get_value() treats non existing paths and existing path with an invalid value
in the same manner, both return None. exits() does an actual dbus request, so
there was no easy way to reduce dbus traffic by not setting a non existing
path. A new function seen is therefore added with returns if there was ever a
succesfull get or change event, so writing to non existing paths can be
prevented in order to reduce dbus traffic.

Perhaps this should always be checked in SetValue?

v2: when requesting the root items the values inside should be marked as seen.