Ensure a bridge conf to prevent Mosquitto crashing
There are two conditions when Mosquitto would crash:

- 'VRM two-way communication' was enabled on installations without
  internet.
- Installations lost their /data partition and then also have 'VRM
  two-way communication' (re)enabled.

Both scenarios create a symlink that is dead, that would only heal
itself once the mosquitto_bridge_registrator.py successfully posted its
auth credentials, which it can't do without internet, or when it it's
reflashed and it tries to overwrite existing credentials.

This fix will create an empty config first, regardless of success.

https://github.com/victronenergy/venus/issues/620