Always restart Mosquitto on new config
The restart responsibility doesn't have to be given to the user of the
class (with the init argument). Because the code is locked with flock,
and a restart is done only when the config changes, only one of the
callers will trigger the needed restart.

This fixes not having a Moquitto bridge when you turn on VRM two-way
communication when there is no internet at first.

NOTE: backwards incompatible change: callers of class must not specify
the restart argument.

https://github.com/victronenergy/venus/issues/642