move existing tree fetching into VeDbusTreeExport
Previously the building of the dictionary that is returned
if you call GetValue or GetText on a non-leaf node was done
in VeDbusService, but only ever called from VeDbusTreeExport.

To make it cleaner, and also easier to add support for GetItems
on the root path (which is a VeDbusTreeExport), move this into
the VeDbusTreeExport class and simply pass a reference to the
VeDbusService that it's maintaining a tree for.