enforce dbus types, when requested.
When adding a dbus path, it is now possible to specify the valuetype.
This will reject writes to the path that is of a type that cannot be
coerced to the right type.

Allowing type-coercion makes us a bit more forgiving when someone
writes an int where a float is expected, or a bool where an int
is expected. Cases that are after all simple enough to handle.