don't invalidate a path prior to deletion
Our dbus-api indicates that you do not have to invalidate a path
before going offline. velib_python has always done this, since
the beginning, and while it didn't hurt in the beginning, it hurts
now. It is also unnecessary and creates additional dbus traffic
for nothing.

The main reason for this change now: Invalidation interferes
with the display of /ProductName and /CustomName on the GUI.
When a service goes offline, it clears the name, leaving the
display blank.