use an arg0namespace for NameOwnerChanges when supported
When subscribing to all NameOwnerChanges, all python script would get
notified twice if a short living script connects to the dbus. This
triggers a lot of unneeded cpu usage.

Since this is a common dbus problem, there is nowaways an arg0namespace;
from the spec: "Match messages whose first argument is of type STRING,
and is a bus name or interface name within the specified namespace. This
is primarily intended for watching name owner changes for a group of
related bus names, rather than for a single name or all name changes.

Since arg0namespace support is new / pending for dbus-python, just try to
use it and if it fails, use the old version.