Convert bridge registrator to FlashMQ
Contrary to what the old comment makes you believe, the mqtt_password file is no
longer just a temporarily backup to make sure the same password is posted in case
that the reply wasn't received, it is actually the leading file now, since the
pasword from the EEPROM gets written to it.

So adjust were to get the password from
  - first the mqtt_password file
  - if that is not there, check the mosquitto config.
  - otherwise generate a random

Don't parse the flashmq config, the password file is leading, not the flashmq
config.