dbusmonitor: explicitly use interface on calls
The dbus protocol allows for omitting the interface in the
header, and then any matching method can be called. This causes
problems with some newer python dbus libraries such as dbus-next
and dbus-fast, and although we patched those to work correctly
in the past, it is better to tighten this down as well.