vedbus: better way for implicit svc registration
Sometimes a service really doesn't need to wait before it registers
on dbus. To accomplish that, we default register to None, and if
that is not overriden in some manner, we do an implicit registration.

If overriden and set to True, then register the service immediately,
but don't log the warning.