refactor, stop using private _locations
This code was using the private _locations list within python dbus
service code, which is just a big no-no.  While there is a
locations property, this is an iterator that will require extra
steps to find the first item. For performance, since registering
the dbus path is the very first thing a service Object does anyway,
just remember the path and use it where required.